Add files via upload
This files works with the Java Version of the SQL*Render program found here: https://github.com/OHDSI/SqlRender
The program assumes there is a SQL Server version of the file in the same directory as the SQL*Render Java file. The name of that files should be naaccr_etl_sqlserver. 
This program will translate the existing file into Redshift, Oracle, and PostgreSQL versions.